Move session_logger tests into existing test module
Consolidate new tests into tests/test_session_logger.py
rather than keeping them in a separate file.
